trash impl

+++ src/api/storage/v1/touch.rs
@@ -29,6 +29,10 @@ async fn touch_task(post: Json<V1PathOnly>) -> Result<V1Response, Box<dyn Error>
         return Err(V1Error::PathOccupied.into());
     }
 
+    if !fs::try_exists(&path_buf.parent().unwrap()).await? {
+        return Err(V1Error::FileNotFound.into())
+    }
+
     fs::File::create(path_buf).await?;
 
     Ok(V1Response::FileItemCreated)
